Split JitManager into two classes
JitManager is now split into two functional elements. The original
class now only looks after the compilation and loading of the library.
The runtime aspects of Jit - allocating memory + wrapping library
functions - have now been moved to RuntimeManager.
